==> 2. ACS Data Collection - Municipalities.R <==
library(tidyverse)
library(dplyr)
library(stringr)
library(sf)
library(modelr)
library(tidycensus)
library(httr)
library(jsonlite)

### Census key

census_api_key("794708f156e5f9a870d2531bb80e777690f72326", install = TRUE)

### Creating filter for municipalities with different minimum wages

municipalities <- c(
  "Alameda city, California",
  "Albuquerque city, New Mexico",
  "Bellingham city, Washington",
  "Belmont city, California",
  "Berkeley city, California",
  "Boulder city, Colorado",
  "Burien city, Washington",
  "Burlingame city, California",
  "Chicago city, Illinois",
  "Cupertino city, California",
  "Daly City city, California",
  "East Palo Alto city, California",
  "Edgewater city, Colorado",
  "Emeryville city, California",
  "Everett city, Washington",
  "Flagstaff city, Arizona",
  "Foster City city, California",
  "Freemont city, California",
  "Half Moon Bay city, California",
  "Hayward city, California",
  "Las Cruces city, New Mexico",
  "Los Angeles city, California",
  "Malibu city, California",
  "Menlo Park city, California",
  "Milpitas city, California",
  "Minneaplois city, Minnesota",
  "Mountain View city, California",
  "Novato city, California",
  "Oakland city, California",
  "Palo Alto city, California",
  "Pasadena city, California",
  "Petaluma city, California",
  "Portland city, Maine",
  "Redwood City city, California",
  "Renton city, Washington",
  "Richmond city, California",
  "Rockland city, Maine",
  "Saint Paul city, Minnesota",
  "San Carlos city, California",
  "San Diego city, California",
  "San Francisco city, California",
  "San Jose city, California",
  "San Leandro city, California",
  "San Mateo city, California",
  "Santa Fe city, New Mexico",
  "Santa Monica city, California",
  "Santa Rosa city, California",
  "Seattle city, Washington",
  "Sonoma city, California",
  "South San Francisco city, California",
  "Sunnyvale city, California",
  "Tucson city, Arizona",
  "Tukwila city, Washington",
  "West Hollywood city, California"
)

### Loading Census Bureau data

# 2010 

acs2010 <- get_acs(survey = "acs1",
                   year = 2010,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     pctlessthanhs = "S1501_C01_002",
                     pcthsgrad = "S1501_C01_003",
                     pctsomecollegeassociates = "S1501_C01_004",
                     pctbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_028",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_015")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = pctlessthanhs / 100,
         prop_hs_grad = pcthsgrad / 100,
         prop_some_college_associates = pctsomecollegeassociates / 100,
         prop_bachelors_higher = pctbachhigher / 100,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2010) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2011

acs2011 <- get_acs(survey = "acs1",
                   year = 2011,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     pctlessthanhs = "S1501_C01_002",
                     pcthsgrad = "S1501_C01_003",
                     pctsomecollegeassociates = "S1501_C01_004",
                     pctbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_028",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_015")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = pctlessthanhs / 100,
         prop_hs_grad = pcthsgrad / 100,
         prop_some_college_associates = pctsomecollegeassociates / 100,
         prop_bachelors_higher = pctbachhigher / 100,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2011) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2012

acs2012 <- get_acs(survey = "acs1",
                   year = 2012,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     pctlessthanhs = "S1501_C01_002",
                     pcthsgrad = "S1501_C01_003",
                     pctsomecollegeassociates = "S1501_C01_004",
                     pctbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_028",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_015")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = pctlessthanhs / 100,
         prop_hs_grad = pcthsgrad / 100,
         prop_some_college_associates = pctsomecollegeassociates / 100,
         prop_bachelors_higher = pctbachhigher / 100,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2012) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2013

acs2013 <- get_acs(survey = "acs1",
                   year = 2013,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     pctlessthanhs = "S1501_C01_002",
                     pcthsgrad = "S1501_C01_003",
                     pctsomecollegeassociates = "S1501_C01_004",
                     pctbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_028",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_015")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = pctlessthanhs / 100,
         prop_hs_grad = pcthsgrad / 100,
         prop_some_college_associates = pctsomecollegeassociates / 100,
         prop_bachelors_higher = pctbachhigher / 100,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2013) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2014

acs2014 <- get_acs(survey = "acs1",
                   year = 2014,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     pctlessthanhs = "S1501_C01_002",
                     pcthsgrad = "S1501_C01_003",
                     pctsomecollegeassociates = "S1501_C01_004",
                     pctbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_028",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_015")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = pctlessthanhs / 100,
         prop_hs_grad = pcthsgrad / 100,
         prop_some_college_associates = pctsomecollegeassociates / 100,
         prop_bachelors_higher = pctbachhigher / 100,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2014) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2015

acs2015 <- get_acs(survey = "acs1",
                   year = 2015,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_028",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2015) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2016

acs2016 <- get_acs(survey = "acs1",
                   year = 2016,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0017",
                     countwhite = "DP05_0032",
                     countblack = "DP05_0033",
                     counthispanic = "DP05_0066",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2016) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2017

acs2017 <- get_acs(survey = "acs1",
                   year = 2017,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0018",
                     countwhite = "DP05_0037",
                     countblack = "DP05_0038",
                     counthispanic = "DP05_0071",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2017) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2018

acs2018 <- get_acs(survey = "acs1",
                   year = 2018,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0018",
                     countwhite = "DP05_0037",
                     countblack = "DP05_0038",
                     counthispanic = "DP05_0071",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2018) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2019

acs2019 <- get_acs(survey = "acs1",
                   year = 2019,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0018",
                     countwhite = "DP05_0037",
                     countblack = "DP05_0038",
                     counthispanic = "DP05_0071",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2019) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2021

acs2021 <- get_acs(survey = "acs1",
                   year = 2021,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0018",
                     countwhite = "DP05_0037",
                     countblack = "DP05_0038",
                     counthispanic = "DP05_0071",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2021) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2022

acs2022 <- get_acs(survey = "acs1",
                   year = 2022,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0018",
                     countwhite = "DP05_0037",
                     countblack = "DP05_0038",
                     counthispanic = "DP05_0071",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2022) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

# 2023

acs2023 <- get_acs(survey = "acs1",
                   year = 2023,
                   geography = "place",
                   variables = c(
                     # gini
                     gini = "B19083_001",
                     # median income by quintile
                     lowest_quintile = "B19081_001",
                     second_quintile = "B19081_002",
                     third_quintile = "B19081_003",
                     fourth_quintile = "B19081_004",
                     highest_quintile = "B19081_005",
                     top_five_percent = "B19081_006",
                     # educational attainment
                     count18to24 = "S1501_C01_001",
                     count24to34 = "S1501_C01_016",
                     count35to44 = "S1501_C01_019",
                     count45to64 = "S1501_C01_022",
                     count65over = "S1501_C01_025",
                     countlessthanhs = "S1501_C01_002",
                     counthsgrad = "S1501_C01_003",
                     countsomecollegeassociates = "S1501_C01_004",
                     countbachhigher = "S1501_C01_005",
                     # total population
                     totalpopulation = "B01003_001",
                     # demographic information
                     medianage = "DP05_0018",
                     countwhite = "DP05_0037",
                     countblack = "DP05_0038",
                     counthispanic = "DP05_0071",
                     # income
                     medianincome = "S1901_C01_012",
                     medianhhincome = "S1901_C02_012",
                     countbelowpoverty = "S1701_C02_001",
                     medianhousingcosts = "S2503_C01_024",
                     # employment
                     pop16plus = "DP03_0001",
                     countlaborforce16plus = "DP03_0002",
                     countunemployedinlaborforce16plus = "DP03_0005",
                     # immigration
                     countforeignborn = "B05002_013", 
                     countforeignborncitizen = "B05002_014",
                     countforeignbornnotcitizen= "B05002_021")) |> 
  select(!moe) |> 
  pivot_wider(names_from = variable,
              values_from = estimate) |> 
  mutate(count18over = count18to24 + count24to34 + count35to44 + count45to64 + count65over,
         prop_less_than_hs = countlessthanhs / count18to24,
         prop_hs_grad = counthsgrad / count18to24,
         prop_some_college_associates = countsomecollegeassociates / count18to24,
         prop_bachelors_higher = countbachhigher / count18to24,
         prop_18_to_24 = count18to24 / totalpopulation,
         prop_24_to_34 = count24to34 / totalpopulation,
         prop_35_to_44 = count35to44 / totalpopulation,
         prop_45_to_64 = count45to64 / totalpopulation,
         prop_65_years_older = count65over / totalpopulation,
         prop_white = countwhite / totalpopulation,
         prop_black = countblack / totalpopulation,
         prop_hispanic = counthispanic / totalpopulation,
         prop_below_poverty = countbelowpoverty / totalpopulation,
         labor_force_part_rate = countlaborforce16plus / pop16plus,
         unemployment_rate = countunemployedinlaborforce16plus / countlaborforce16plus,
         prop_foreign_born = countforeignborn / totalpopulation,
         prop_foreign_born_undocumented = countforeignbornnotcitizen / totalpopulation,
         year = 2023) |> 
  rename(median_age = medianage,
         median_income = medianincome, 
         median_housing_costs = medianhousingcosts,
         total_population = totalpopulation,
         geoid = GEOID,
         name = NAME) |> 
  select(geoid,
         name,
         year,
         gini,
         lowest_quintile,
         second_quintile,
         third_quintile,
         fourth_quintile,
         highest_quintile,
         top_five_percent,
         total_population,
         prop_less_than_hs,
         prop_hs_grad,
         prop_some_college_associates,
         prop_bachelors_higher,
         prop_18_to_24,
         prop_24_to_34,
         prop_35_to_44,
         prop_45_to_64,
         prop_65_years_older,
         prop_white,
         prop_black,
         prop_hispanic,
         prop_below_poverty,
         labor_force_part_rate,
         unemployment_rate,
         median_age,
         median_income,
         median_housing_costs,
         prop_foreign_born,
         prop_foreign_born_undocumented)

### Binding ACS Data Together 

acs_binded_cities <- rbind(acs2010, acs2011, acs2012, acs2013, acs2014, acs2015, acs2016,
                    acs2017, acs2018, acs2019, acs2021, acs2022, acs2023) |> 
  rename(statecountycity_fips = geoid) |> 
  mutate(prop_less_than_hs = as.numeric(prop_less_than_hs),
         prop_hs_grad = as.numeric(prop_hs_grad),
         prop_some_college_associates = as.numeric(prop_some_college_associates),
         prop_bachelors_higher = as.numeric(prop_bachelors_higher),
         prop_18_to_24 = as.numeric(prop_18_to_24),
         prop_24_to_34 = as.numeric(prop_24_to_34),
         prop_35_to_44 = as.numeric(prop_35_to_44),
         prop_45_to_64 = as.numeric(prop_45_to_64),
         prop_65_years_older = as.numeric(prop_65_years_older),
         prop_white = as.numeric(prop_white),
         prop_black = as.numeric(prop_black),
         prop_hispanic = as.numeric(prop_hispanic),
         labor_force_part_rate = as.numeric(labor_force_part_rate),
         unemployment_rate = as.numeric(unemployment_rate),
         prop_foreign_born = as.numeric(prop_foreign_born),
         prop_foreign_born_undocumented = as.numeric(prop_foreign_born_undocumented),
         statecounty_fips = substr(statecountycity_fips, 1, 5),
         state_fips = substr(statecountycity_fips, 1, 2)) |> 
  select(state_fips, statecounty_fips, everything())

### Removing unnecessary dataframes

rm(list = ls()[grepl("acs20", ls())])

### Saving dataframe

save.image(here("Processed/city_acs.RData"))
